requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.42.0
django-debug-toolbar==4.2.0
django-redis==5.4.0
exceptiongroup==1.2.0
executing==2.0.1
fake-useragent==1.5.0
frozenlist==1.4.1
idna==3.6
importlib_resources==6.3.0
ipython==8.18.1
jedi==0.19.1
lxml==5.1.0
matplotlib-inline==0.1.6
mimetypes-extensions==0.1.0
multidict==6.0.5
oauthlib==3.2.2
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
stripe==8.6.0
traitlets==5.14.1
twilio==9.0.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
zipp==3.18.0
